requirements.txt
asgiref==3.8.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.41.0
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.10.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-environ==0.12.0
django-extensions==4.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.3.0
psycopg==3.2.4
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pymongo==3.11.4
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.5.0
